refactor(aw): trim graft observability dependency edge
Finding: AW-READY-O7 (C15)
