+++ boundaries/atm-graft-python/hermes-graft-binding.toml
@@ -22,7 +22,7 @@ io_forbidden = ["database_io", "direct_socket_io", "recipient_routing"]
 
 [dependencies]
 allowed_dependents = []
-allowed_dependencies = ["atm-core", "atm-graft", "atm-observability", "pydantic", "serde_json"]
+allowed_dependencies = ["atm-core", "atm-graft", "pydantic", "serde_json"]
 forbidden_edges = [
   "atm-graft-python -> atm-daemon",
   "atm-graft-python -> atm-storage-rusqlite",
