+++ crates/atm-core/src/observability.rs
@@ -1,7 +1,7 @@
 //! ATM-owned observability boundary and projected log/health types.
 
 use std::fmt;
-use std::path::PathBuf;
+use std::path::{Path, PathBuf};
 
 use sc_lint_attributes::sc_lint;
 use serde::de::Error as DeError;
@@ -15,6 +15,42 @@ use crate::observability_counters::{JsonlDiagnosticCounters, TimelineDiagnosticC
 use crate::protocol::RequestId;
 use crate::schema::AtmMessageId;
 use crate::types::{AgentName, IsoTimestamp, TaskId, TeamName};
+
+/// Canonical retained event file shared by daemon and CLI projections.
+pub const CANONICAL_LOG_FILE_NAME: &str = "atm.log.jsonl";
+/// AW.4's dedicated graft fallback satellite file.
+pub const GRAFT_FALLBACK_LOG_FILE_NAME: &str = "atm-graft-fallback.jsonl";
+
+/// The redaction boundary shared by retained-log producers.
+pub const RETAINED_FIELD_ALLOWLIST: &[&str] = &[
+    "ts",
+    "level",
+    "component",
+    "code",
+    "command",
+    "action",
+    "correlation_id",
+    "outcome",
+    "elapsed_ms",
+    "attempt",
+    "strategy",
+    "endpoint_kind",
+    "failure_class",
+    "refresh_error_code",
+    "error_layer",
+    "origin",
+];
+
+/// Removes fields that are not permitted to reach retained diagnostics.
+pub fn sanitize_retained_fields(mut fields: Map<String, Value>) -> Map<String, Value> {
+    fields.retain(|key, _| RETAINED_FIELD_ALLOWLIST.contains(&key.as_str()));
+    fields
+}
+
+/// Returns the dedicated graft fallback satellite path.
+pub fn graft_fallback_log_path(log_dir: &Path) -> PathBuf {
+    log_dir.join(GRAFT_FALLBACK_LOG_FILE_NAME)
+}
 
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq, Hash)]
 #[serde(transparent)]
