+++ crates/atm-graft-python/Cargo.toml
@@ -29,7 +29,6 @@ extension-module = ["pyo3/extension-module"]
 [dependencies]
 atm-graft.workspace = true
 atm-core.workspace = true
-atm-observability.workspace = true
 pyo3 = "0.29"
 serde_json.workspace = true
 tokio.workspace = true
