+++ docs/plans/phase-aw/phase-aw-plan.md
@@ -28,8 +28,8 @@ dependency_relations:
   - prerequisite: AW.1
     dependent: AW.4
     relation: must_follow
-    scope: AW.4 consumes AW.1's graft_fallback_log_path constant and the
-      atm-observability boundary record AW.1 creates
+    scope: AW.4 consumes AW.1's graft_fallback_log_path contract from
+      atm-core and the retained-log boundary record AW.1 creates
   - prerequisite: AW.2
     dependent: AW.3
     relation: must_follow
@@ -333,7 +333,7 @@ addressed in round 3 text):
 | I7 | `DiagnosticSink` never specified | Fixed: signature in §2 "Cross-sprint contracts" |
 | I8 | AW.3/AW.5 file overlap not reserved | Fixed: §4 file-reservation table |
 | I9 | AW.1 D6 hedge; AW.3 health block crossed an unrecorded crate edge | Fixed: D6 lists exact contents; AW.3 uses `DiagnosticCountersSource` injected by bootstrap and reaches the timeline via `atm-runtime` state — no new edge |
-| I10 | `atm-graft-python` grant in tracing-bridge.toml broader than needed | Fixed: AW.1 grants `allowed_dependents = ["atm-daemon-bootstrap", "atm"]` only; AW.4 amends the record when it adds the edge |
+| I10 | `atm-graft-python` grant in tracing-bridge.toml broader than needed | Fixed: the Python binding consumes the retained contract from `atm-core` and has no tracing-bridge dependency |
 | I11 | ATM-QA-006 — AW.5 D2 claimed `atm_ack` parity without a test or AC | Fixed: AW.5 D3 parity test adds the ack case; AC6 asserts `atm_ack` equals `atm ack --json` |
 | I12 | RBQA-AW5-F002 — ack result type unnamed, no `response_types` entry | Fixed: `AtmAckResult` (= `AtmSendResult`) named in AW.4 D3a and committed in AW.4 D5; AW.5 D4 re-asserts the full list |
 | M13 | Minor — `docs/requirements.md:907` stale version | Fixed in AW.1 D5 (see I6) |
